requirements.txt
pydantic
fastapi
uvicorn
sentence-transformers
rank-bm25
scikit-learn
PyYAML
pytest
